+++ virtinst/diskbackend.py
@@ -391,7 +391,7 @@ class CloneStorageCreator(_StorageCreator):
         else:
             vfs = os.statvfs(os.path.dirname(self._path))
             avail = vfs.f_frsize * vfs.f_bavail
-        need = long(self._size * 1024 * 1024 * 1024)
+            need = int(self._size) * 1024 * 1024 * 1024
         if need > avail:
             if self._sparse:
                 msg = _("The filesystem will not have enough free space"
@@ -411,7 +411,7 @@ class CloneStorageCreator(_StorageCreator):
         text = (_("Cloning %(srcfile)s") %
                 {'srcfile': os.path.basename(self._input_path)})
 
-        size_bytes = long(self.get_size() * 1024 * 1024 * 1024)
+        size_bytes = int(self.get_size()) * 1024 * 1024 * 1024
         progresscb.start(filename=self._output_path, size=size_bytes,
                          text=text)
